Refactor control flow examples to define blocks directly in workflows
- Removed utility functions for creating blocks and defined blocks directly within each control flow example (branch, parallel, sequential, do-until, do-while, foreach).
- Enhanced clarity and maintainability by consolidating block definitions with their respective workflows.
- Updated logging for block execution to ensure consistent output across examples.
