+++ packages/cue/src/examples/control-flows/sequential.ts
@@ -1,11 +1,31 @@
 import { Cue } from '../../../src/cue';
 import { z } from 'zod';
 import { log, monitorCue } from './utils/logger';
-import { createMathBlocks } from './utils/blocks';
 
 // Create sequential workflow
 const createSequentialWorkflow = () => {
-  const { addBlock, multiplyBlock } = createMathBlocks();
+  // Define blocks
+  const addBlock = Cue.createBlock({
+    id: 'add',
+    inputSchema: z.object({ a: z.number(), b: z.number() }),
+    outputSchema: z.number(),
+    execute: async ({ inputData }) => {
+      const result = inputData.a + inputData.b;
+      log('Add block executed', { input: inputData, result });
+      return result;
+    },
+  });
+
+  const multiplyBlock = Cue.createBlock({
+    id: 'multiply',
+    inputSchema: z.number(),
+    outputSchema: z.number(),
+    execute: async ({ inputData }) => {
+      const result = inputData * 2;
+      log('Multiply block executed', { input: inputData, result });
+      return result;
+    },
+  });
 
   const cue = Cue.createCue({
     id: 'sequential-workflow',
